backend/pyproject.toml
[tool.poetry]
name = "workmate-test-task"
version = "0.1.0"
description = ""
authors = ["Artem Yurkov <yurkov.artyom21@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
django-colorfield = "^0.11.0"
django-filter = "^24.3"
gunicorn = "^23.0.0"
drf-yasg = "^1.21.7"
setuptools = "^75.1.0"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
requests = "^2.32.3"
django = "4.2.16"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
